+++ packages/contract_sdk/lib/signaling_contract_sdk.dart
@@ -5,4 +5,11 @@
 library signaling_contract_sdk;
 
 export 'generated/contracts.dart';
-export 'signaling_contract_extensions.dart';
+export 'signaling_contract_extensions.dart'
+    show
+        SignalingDataCompression,
+        CompressibleData,
+        StringData,
+        BytesData,
+        IntListData,
+        SignalingContractExtension;